refactor(core): move functionality from Dev/System to Core
