+++ modules/admin/core/src/core-routing.module.ts
@@ -1,9 +1,8 @@
 import { NgModule } from '@angular/core'
 import { Routes, RouterModule } from '@angular/router'
 
-import { AboutComponent } from './components/about.component'
-import { DashboardComponent } from './components/dashboard.component'
 import { IndexComponent } from './components/index.component'
+import { InfoComponent } from './components/info.component'
 
 const routes: Routes = [ {
   path: '',
@@ -12,9 +11,9 @@ const routes: Routes = [ {
   },
   children: [
     { path: '', component: IndexComponent, children: [
-      { path: '', redirectTo: 'index', pathMatch: 'full' },
-      { path: 'index', component: DashboardComponent, data: { title: 'Dashboard' } },
-      { path: 'about', component: AboutComponent, data: { title: 'About' } },
+      { path: '', redirectTo: 'info', pathMatch: 'full' },
+      { path: 'info', component: InfoComponent },
+      { path: 'modules', loadChildren: './modules/modules.module#ModulesModule' },
     ]}
   ]
 } ]
