+++ modules/admin/core/src/core.module.ts
@@ -1,17 +1,14 @@
 import { NgModule } from '@angular/core'
-import { CommonModule } from '@angular/common'
 import { ColmenaUiModule } from '@colmena/admin-ui'
 
 import { CoreRoutingModule } from './core-routing.module'
 import { CoreService } from './core.service'
 
-import { AboutComponent } from './components/about.component'
-import { DashboardComponent } from './components/dashboard.component'
 import { IndexComponent } from './components/index.component'
+import { InfoComponent } from './components/info.component'
 
 @NgModule({
   imports: [
-    CommonModule,
     ColmenaUiModule,
     CoreRoutingModule,
   ],
@@ -19,9 +16,8 @@ import { IndexComponent } from './components/index.component'
     CoreService,
   ],
   declarations: [
-    AboutComponent,
-    DashboardComponent,
     IndexComponent,
+    InfoComponent,
   ],
 })
 export class CoreModule { }
