--- modules/admin/dev/src/colmena/components/dashboard.component.ts
@@ -1,6 +1,0 @@
-import { Component } from '@angular/core'
-
-@Component({
-  template: `dashboard`
-})
-export class DashboardComponent {}
