--- modules/admin/dev/src/colmena/components/data.component.ts
@@ -1,6 +1,0 @@
-import { Component } from '@angular/core'
-
-@Component({
-  template: `data`
-})
-export class DataComponent {}
