--- modules/admin/dev/src/colmena/components/detail.component.ts
@@ -1,6 +1,0 @@
-import { Component } from '@angular/core'
-
-@Component({
-  template: `Detail`
-})
-export class DetailComponent {}
