+++ modules/admin/dev/src/components/index.component.ts
