+++ modules/admin/system/src/system.module.ts
@@ -2,8 +2,6 @@ import { NgModule } from '@angular/core'
 import { ColmenaUiModule } from '@colmena/admin-ui'
 
 import { SystemDashboardComponent } from './system.component'
-import { SystemInfoComponent } from './system-info.component'
-import { SystemModulesComponent } from './system-modules.component'
 
 import { SystemRoutingModule } from './system-routing.module'
 import { HasSystemAccess } from './system.guards'
@@ -15,8 +13,6 @@ import { HasSystemAccess } from './system.guards'
   ],
   declarations: [
     SystemDashboardComponent,
-    SystemInfoComponent,
-    SystemModulesComponent,
   ],
   providers: [
     HasSystemAccess,
