+++ modules/api/core/common/models/core.js
@@ -1,6 +1,11 @@
 'use strict'
 
+const { getAppConfigs } = require('@colmena/api-helpers')
+
 module.exports = function(Core) {
+  Core.ping = () => Core.app.models.Ping.ping()
+
+  Core.modules = () => Promise.resolve(getAppConfigs())
   // This is a static method that lives on the Core model
   Core.echo = value => Promise.resolve(value)
