+++ modules/api/core/common/models/core.json
@@ -2,5 +2,53 @@
   "name": "Core",
   "plural": "Core",
   "description": "Core: basic model",
-  "base": "Model"
+  "base": "Model",
+  "acls": [
+    {
+      "accessType": "EXECUTE",
+      "principalType": "ROLE",
+      "principalId": "$everyone",
+      "permission": "DENY",
+      "property": [
+        "modules",
+        "ping"
+      ]
+    },
+    {
+      "accessType": "EXECUTE",
+      "principalType": "ROLE",
+      "principalId": "system-admin",
+      "permission": "ALLOW",
+      "property": [
+        "modules",
+        "ping"
+      ]
+    }
+  ],
+  "methods": {
+    "modules": {
+      "returns": {
+        "arg": "result",
+        "type": "Object",
+        "root": true
+      },
+      "description": "Show the modules loaded by the API",
+      "http": {
+        "path": "/modules",
+        "verb": "get"
+      }
+    },
+    "ping": {
+      "returns": {
+        "arg": "result",
+        "type": "Object",
+        "root": true
+      },
+      "description": "Ping the API to retrieve its internal state",
+      "http": {
+        "path": "/ping",
+        "verb": "get"
+      }
+    }
+  }
 }
