+++ modules/api/core/package.json
@@ -7,7 +7,9 @@
   "license": "MIT",
   "author": "Bram Borggreve <bram@colmena.io> (https://github.com/beeman) (https://twitter.com/beeman_nl)",
   "url": "https://github.com/colmena/colmena",
-  "dependencies": {},
+  "dependencies": {
+    "@colmena/api-helpers": "^0.1.0"
+  },
   "colmena": {
     "module": {
       "modelSources": [
